moved security related use cases in securityproblems package
