+++ app/src/main/java/com/babylon/wallet/android/presentation/settings/securitycenter/SecurityCenterViewModel.kt
@@ -3,7 +3,7 @@ package com.babylon.wallet.android.presentation.settings.securitycenter
 import androidx.lifecycle.viewModelScope
 import com.babylon.wallet.android.di.coroutines.DefaultDispatcher
 import com.babylon.wallet.android.domain.model.SecurityProblem
-import com.babylon.wallet.android.domain.usecases.GetSecurityProblemsUseCase
+import com.babylon.wallet.android.domain.usecases.securityproblems.GetSecurityProblemsUseCase
 import com.babylon.wallet.android.presentation.common.StateViewModel
 import com.babylon.wallet.android.presentation.common.UiState
 import dagger.hilt.android.lifecycle.HiltViewModel
