+++ app/src/main/java/com/babylon/wallet/android/presentation/settings/securitycenter/securityfactors/SecurityFactorsViewModel.kt
@@ -1,7 +1,8 @@
 package com.babylon.wallet.android.presentation.settings.securitycenter.securityfactors
 
 import androidx.lifecycle.viewModelScope
-import com.babylon.wallet.android.domain.usecases.GetSecurityProblemsUseCase
+import com.babylon.wallet.android.di.coroutines.DefaultDispatcher
+import com.babylon.wallet.android.domain.usecases.securityproblems.GetSecurityProblemsUseCase
 import com.babylon.wallet.android.presentation.common.StateViewModel
 import com.babylon.wallet.android.presentation.common.UiState
 import com.babylon.wallet.android.presentation.settings.SettingsItem
