+++ app/src/main/java/com/babylon/wallet/android/presentation/settings/troubleshooting/reset/FactoryResetViewModel.kt
@@ -3,7 +3,7 @@ package com.babylon.wallet.android.presentation.settings.troubleshooting.reset
 import androidx.lifecycle.viewModelScope
 import com.babylon.wallet.android.domain.model.SecurityProblem
 import com.babylon.wallet.android.domain.usecases.DeleteWalletUseCase
-import com.babylon.wallet.android.domain.usecases.GetSecurityProblemsUseCase
+import com.babylon.wallet.android.domain.usecases.securityproblems.GetSecurityProblemsUseCase
 import com.babylon.wallet.android.presentation.common.OneOffEvent
 import com.babylon.wallet.android.presentation.common.OneOffEventHandler
 import com.babylon.wallet.android.presentation.common.OneOffEventHandlerImpl
